Значения аттрибутов в индексе elastic синхронизированы с таблицей в БД (CDU)
