+++ project/src/main/java/caselab/elastic/ElasticSearchInterface.java
@@ -1,0 +1,7 @@
+package caselab.elastic;
+
+import org.springframework.data.domain.Page;
+
+public interface ElasticSearchInterface<T> {
+    Page<T> searchValuesElastic(String searchText, int page, int size);
+}
